requirements/base.txt
# ── Core framework ────────────────────────────────────────────────────────────
langgraph
langchain-core
langchain-groq
python-dotenv
typing_extensions
pydantic>=2.6.0

# ── LLM ───────────────────────────────────────────────────────────────────────
groq

# ── Database ──────────────────────────────────────────────────────────────────
psycopg2-binary

# ── HTTP ──────────────────────────────────────────────────────────────────────
requests
